The Variable needs time bounds.
This is just a fake time to make the ILAMB analysis run. My analysis expects temporal variables, even if they are defined in just one time slice. So I had you add this fake time. But when I compute the temporal mean, I use the time bounds to determine how long the interval was. In the case of a single time, I have no way to estimate how long that time was, so it is left to the single value you gave me.

So, when I took the temporal mean in the analysis phase, it was all 0. Then ILAMB looks to prune away regions where there is no data and we skip those in the plotting / html generation phase. Since there was no data on the global region, it was skipped.