Added to avoid underflow errors in masked arrays.
This bit of logic opens up a block where all underflow errors will be ignored, and then resets to whatever it was once the block is exitted. We have to do this a lot because netCDF gives us masked arrays where the masked value is typically something terrible for floating point arithmetic, like 1e32. The model you ran may have not given you this trouble, but others will.